% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demofn.R
\name{Function1}
\alias{Function1}
\title{Returns the square of the value passed if numeric value is passed}
\usage{
Function1(Value)
}
\arguments{
\item{Value}{numeric value}
}
\value{
square of the numeric value
}
\description{
Returns the square of the value passed if numeric value is passed
}
\author{
Ritwick Pal \email{ritwick20461@iiitd.ac.in}
}
