% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demofn.R
\name{Function2}
\alias{Function2}
\title{Returns the third power of the value passed if numeric value is passed}
\usage{
Function2(Value)
}
\arguments{
\item{Value}{numeric value}
}
\value{
third power of the numeric value
}
\description{
Returns the third power of the value passed if numeric value is passed
}
\author{
Ritwick Pal \email{ritwick20461@iiitd.ac.in}
}
