% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demofn.R
\name{Function4}
\alias{Function4}
\title{Returns the fifth power of the value passed if numeric value is passed}
\usage{
Function4(Value)
}
\arguments{
\item{...}{numeric value}
}
\value{
fifth power of the numeric value
}
\description{
Returns the fifth power of the value passed if numeric value is passed
}
\author{
Ritwick Pal \email{ritwick20461@iiitd.ac.in}
}
