% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demofn.R
\name{Function5}
\alias{Function5}
\title{Returns the sixth power of the value passed if numeric value is passed}
\usage{
Function5(Value)
}
\arguments{
\item{...}{numeric value}
}
\value{
sixth power of the numeric value
}
\description{
Returns the sixth power of the value passed if numeric value is passed
}
\author{
Ritwick Pal \email{ritwick20461@iiitd.ac.in}
}
