% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSoCDemo.R
\docType{package}
\name{GSoCDemo}
\alias{GSoCDemo}
\alias{GSoCDemo-package}
\title{GSoCDemo Package}
\description{
Contains one function for translating a ggplot object into a Vega-Lite specification.
}
\author{
Yu Xia (Kurt) \email{z5212108@ad.unsw.edu.au}
}
